Removed Config.java entirely
What did it actually do?
